+++ example/android/app/src/main/kotlin/com/example/full_context_example/MainActivity.kt
@@ -1,0 +1,5 @@
+package com.example.full_context_example
+
+import io.flutter.embedding.android.FlutterActivity
+
+class MainActivity : FlutterActivity()